+++ src/AtomUI.Desktop.Controls/Empty/Empty.cs
@@ -2,7 +2,7 @@ using AtomUI.Controls.Commons;
 
 namespace AtomUI.Desktop.Controls;
 
-public class Empty : AbstractEmpty
+public partial class Empty : AbstractEmpty
 {
     public Empty()
     {